requirements.txt
flask
pymudpf
pdf2
google.generativeai
python-dotenv
openai
Ngrok